Home page lists now built from DB, moved Spotify API + DB data to external files
So there's now a DB with a table called "Playlists" and four rows: playlist_name, playlist_spotify_id, playlist_icon_name, playlist_bg_name. This is used to build up the front page and links to the genre lists.

I've also moved the Spotify API credentials and Database credentials, to external files. Now you just have to set them to your custom values and the site does the rest of the work.