Added Error Handling, Added favicon
The error handling is a bit crude, but I believe it handles the majority of common incidents. I could have just try-catched the whole of the data receiving in the Angular, but it didn't seem right.

Error reporting is a simple request to e-mail the admin. I could have made it log errors, but I don't expect this to get too big, so it's not really a necessary addition, at this point.