Split up Angular code, and minified
Angular code sections are now in separate files. I probably could have taken out some functions from the genreController and thrown them in factories, but they were too small for me to deem it worth it.

Altered code to work for minification (forgot to do that when I was first writing it), and minified.

Considered using RequireJS for the files I split up, but I always find that library a little unintuitive and annoying to use, so I've excluded it in this project. Maybe if the project were bigger.